# Use generated unmarshalers for some common proto messages.