--- a/proto/data.proto
+++ b/proto/data.proto
@@ -24,6 +24,7 @@ import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 // Timestamp represents a state of the hybrid logical clock.
 message Timestamp {
   option (gogoproto.goproto_stringer) = false;
+  option (gogoproto.unmarshaler) = true;
 
   // Holds a wall time, typically a unix epoch time
   // expressed in nanoseconds.
@@ -40,6 +41,8 @@ message Timestamp {
 // basic types: a "bag o' bytes" generic byte slice and an incrementable
 // int64, for use with the Increment API call.
 message Value {
+  option (gogoproto.unmarshaler) = true;
+
   // Bytes is the byte slice value. If this field is set, the integer field
   // should not be.
   optional bytes bytes = 1;
@@ -65,6 +68,8 @@ message Value {
 // MVCCValue differentiates between normal versioned values and
 // deletion tombstones.
 message MVCCValue {
+  option (gogoproto.unmarshaler) = true;
+
   // True to indicate a deletion tombstone. If false, value should not
   // be nil.
   optional bool deleted = 1 [(gogoproto.nullable) = false];
@@ -173,6 +178,8 @@ enum TransactionStatus {
 // NodeList keeps a growing set of NodeIDs as a sorted slice, with Add()
 // adding to the set and Contains() verifying membership.
 message NodeList {
+  option (gogoproto.unmarshaler) = true;
+
   // Note that this does not use the NodeID custom type because that appears
   // to interact badly with the repeated and/or packed options.
   repeated int32 nodes = 1 [packed=true];
@@ -186,6 +193,7 @@ message NodeList {
 // contention.
 message Transaction {
   option (gogoproto.goproto_stringer) = false;
+  option (gogoproto.unmarshaler) = true;
 
   optional string name = 1 [(gogoproto.nullable) = false];
   // Key is the key which anchors the transaction. This is typically
@@ -239,6 +247,8 @@ message Transaction {
 
 // MVCCMetadata holds MVCC metadata for a key. Used by storage/engine/mvcc.go.
 message MVCCMetadata {
+  option (gogoproto.unmarshaler) = true;
+
   optional Transaction txn = 1;
   // The timestamp of the most recent versioned value.
   optional Timestamp timestamp = 2 [(gogoproto.nullable) = false];
@@ -259,6 +269,8 @@ message MVCCMetadata {
 // GCMetadata holds information about the last complete key/value
 // garbage collection scan of a range.
 message GCMetadata {
+  option (gogoproto.unmarshaler) = true;
+
   // The last GC scan timestamp in nanoseconds since the Unix epoch.
   optional int64 last_scan_nanos = 1 [(gogoproto.nullable) = false];
   // The oldest unresolved write intent in nanoseconds since epoch.
@@ -269,15 +281,17 @@ message GCMetadata {
 // TimeSeriesDatapoint is a single point of time series data; a value associated
 // with a timestamp.
 message TimeSeriesDatapoint {
-    // The timestamp when this datapoint is located, expressed in nanoseconds
-    // since the unix epoch.
-    optional int64 timestamp_nanos = 1 [(gogoproto.nullable) = false];
-    // An integer representation of the value of this datapoint. If this field
-    // is set, then 'float_value' must not be set.
-    optional int64 int_value = 2;
-    // A floating point representation of the value of this datapoint. If this
-    // field is set, then 'int_value' must not be set.
-    optional float float_value = 3;
+  option (gogoproto.unmarshaler) = true;
+
+  // The timestamp when this datapoint is located, expressed in nanoseconds
+  // since the unix epoch.
+  optional int64 timestamp_nanos = 1 [(gogoproto.nullable) = false];
+  // An integer representation of the value of this datapoint. If this field
+  // is set, then 'float_value' must not be set.
+  optional int64 int_value = 2;
+  // A floating point representation of the value of this datapoint. If this
+  // field is set, then 'int_value' must not be set.
+  optional float float_value = 3;
 }
 
 // TimeSeriesData is a set of observations of a single variable value at
@@ -285,9 +299,11 @@ message TimeSeriesDatapoint {
 // identifies the source variable, and a repeated set of TimeSeriesDatapoint
 // messages representing distinct measurements of that variable.
 message TimeSeriesData {
-    // A string which uniquely identifies the variable from which this data was
-    // measured.
-    optional string name = 1 [(gogoproto.nullable) = false];
-    // Datapoints representing one or more measurements taken from the variable.
-    repeated TimeSeriesDatapoint datapoints = 2;
+  option (gogoproto.unmarshaler) = true;
+
+  // A string which uniquely identifies the variable from which this data was
+  // measured.
+  optional string name = 1 [(gogoproto.nullable) = false];
+  // Datapoints representing one or more measurements taken from the variable.
+  repeated TimeSeriesDatapoint datapoints = 2;
 }