… proto/data.pb.go
… storage/engine/data.pb.cc