The generated unmarshalers are a lot faster and unmarshaling was a
significant CPU consumer in the MVCCGet benchmark. The MVCCScan
benchmarks showed similar speedups.

benchmark                       old ns/op     new ns/op     delta
BenchmarkMVCCGet1Version        4021          3572          -11.17%
BenchmarkMVCCGet10Versions      4432          3818          -13.85%
BenchmarkMVCCGet100Versions     5236          4680          -10.62%

benchmark                       old MB/s     new MB/s     speedup
BenchmarkMVCCGet1Version        254.60       286.64       1.13x
BenchmarkMVCCGet10Versions      231.00       268.14       1.16x
BenchmarkMVCCGet100Versions     195.55       218.77       1.12x